Bug fix for USB debugger not connecting later.
The USB debugger wouldn't connect unless you got it on the initial
breakpoint. This turned out to be a bad combination of the 5-second
probing interval, the 2-second listening interval, and the incorrectly set
USB toggle.

Each time KdpConnect is called, the FTDI device is reset. It was
incorrectly clearing its toggles on the bulk endpoints. So an initial
transmission right after a reset was potentially getting thrown out if the
toggles were wrong. Kd bailed out after 2 seconds, and the synchronize
probe was sent again after 5, not seeing the initial synchronize from Kd.
This went on indefinitely.