Make stderr unbuffered by default.
ssh-keygen was using fprintf to print a prompt to stderr without a new
line. Since stderr was line buffered, the prompt was not printing,
creating user confusion. While ssh-keygen probably shouldn't be printing a
prompt that requires user interaction to stderr, the rest of the OS world
defaults to have stderr unbuffered. This change puts Minoca in line with
the others. That said, POSIX only specifies that stderr should not be fully
buffered, but not whether it should be line buffered or unbuffered.

This change also adds an optimization for fprintf's to unbuffered streams.
It will cache the written characters in a stack allocated buffer and then
flush them to the unbuffered stream when the print is done. This saves a
lot of write system calls.