Prevent Netlink from receiving physical packets.
The Netlink network entry should not be receiving packets that come up
from the physical layer. It exists as a network to be the bottom layer of
network messages, routing those messages to the correct sockets. It has no
physical layer below it. It was, however, advertising its Ethernet
protocol number as 0, causing it to match any uninitialized packets that
the physical layer reported.

This change introduces an invalid protocol number that non-physical
networks can use to prevent being invoked by uninitialized packets.