Fix PCNet32 link change detection on VMWare.
There isn't a very good story for link change detection for PCNet32
devices that lack a PHY. The LED status register seemed like the best bet
(as that's what QEMU and VirtualBox toggle on link up/down transitions).
That said, VMWare appears to handle it differently, and actually do a
better job of being in line with the hardware spec when it comes to
full-duplex mode. This change fixes link detection for VMWare by setting
some extra bits and checking the LED status register differently.