Fix hangs during third-party build on Windows.
On only my Windows 10 machine, the build would seem to hang sometimes, but
then keep going when I gave focus back to the build window. Whenever I
attached a debugger, it seemed to be stuck in ShCollectOutput. I think
the thread returned by _beginthread was exiting too quickly, and then
some other handle got waited on in WaitForSingleObject. That handle
happened to signal when the window regained focus.

The fix is to use _beginthreadex, since that handle doesn't go away until
it's explictly closed.