DNS fixes, including support for SOA records.
DNS was failing on the Type 0 machine we're trying to get up and running,
because the DNS server was returning an SOA record we didn't know what to
do with. We also didn't have support for finding the address of a returned
DNS server, which is needed in order to support these SOA records.