Leave owner permission bits alone for .ssh files.
The recent change in the init.sh script was setting the permissions of the
files and directories checked by the SSH server to 700.

This change switches that to flipping off all bits in group and other, but
leaving the user bits alone. This way if the user wants to keep her id_rsa
at 400, she's not fighting with this script.