Reap e100 transmit commands every half ring cycle.
This change makes the e100 driver set the interrupt bit on a transmit
command if only half of the available command slots are free. This will
force the hardware to interrupt and the driver to reap commands after half
the ring is processed during heavy traffic. The driver used to only reap
commands when the hardware went into suspension, meaning that pending
packets could have had to wait for the full ring to be processed before
getting a chance to run.