Force e100 to resume command unit on send.
The e100 command unit should not need to be resumed unless it is in the
suspend state. There is, however, a bug in QEMU where it stops processing
commands after encountering 16 in a row, but fails to put the transmit
command unit into the suspended state. This left the command unit active
while unsent network packets were in the command queue. Future attempts to
resend the packets were not sending the resume command to kickstart the
command unit because it was claiming to be active. This change forces the
sending of the resume command when new packets are added to the command
queue.