Clear KPROCESS.BinaryName when Environment is cleared.
The BinaryName field points to a buffer within the environment, so it must
be cleared out when the environment is destroyed. Before 'ps' could run
and access some random heap area.

I also hit a busted process child list: the list entry on the list had a
NULL next pointer. My theory is that this is somehow caused by a
second queuing of the child signal after the process has terminated. I've
created an assert to hopefully catch that case.