Fix extra TCP timer reference.
If a connect fails, we transition back from a syn-sent or syn-received
state back to initialized, so that connect can try again. But we fail to
release the timer reference, so upon closing we end up with an extra
reference.

Also added a PreviousState member of the TCP socket, for easier debugging.