Enable SA_RESTART for connect().
Handling STATUS_INTERRUPTED at the end of IoSysSocketConnect got left out
when the SA_RESTART support was added. Unfortunately, the fix to enable
connect to restart needed a bit more work than just checking the return
value in IoSysSocketConnect.

An interrupted connect leaves the TCP state machine running in the
background, which means that the TCP socket's state may be something other
than 'Initialized' when NetpTcpConnect gets called by the SA_RESTART
mechanism. But calling connect() on a socket that is already connecting or
already establish is suppoed to return an error. This change bypasses those
errors for the restart case by recording when a socket's connect gets
interrupted and allowing one call to NetpTcpConnect when the socket is not
in the 'Initialized' state. And since the TCP state machine is already in
some more advanced state, connect just needs to wait on the I/O state,
rather than sending a SYN.