Don't leak KSTATUS to C library via SO_ERROR.
The SO_ERROR socket option returns (and clears) the last socket error.
Naturally, the core networking driver was saving and returning a KSTATUS
for the last error value. The C library, however, was not converting this
to a POSIX error code. Add the conversion, but be careful about the size
of the buffer returned by the kernel - it copies the error value, but will
blindly truncate the LONG KSTATUS if the buffer is smaller.