Fix AHCI enumerate when no media is present.
If there are no AHCI devices present, then an empty child array was being
passed to IoMergeChildArrays. Most callers of IoMergeChildArrays check for
a 0 child count, so this change makes AchipEnumeratePorts follow suit. It
also hardens IoMergeChildArrays so that it will not call MmAllocatePool
with a size of 0 in the case where an empty array is supplied.

This bug surfaces on the Gizmo 2.