Add _OSI to ACPI.
Found some new bugs on a Type 1 Packet machine. The primary one was
missing support for the _OSI method, which is supposed to pre-exist in the
namespace.

Also enabled stores for a package clobbering a package, since that
firmware seemed to be creating a package with empty packages inside, and
then clobbering those empty packages with fuller ones later.