Fix AHCI finding its proper BAR.
The AHCI BAR is a a specific one in PCI config space. This is so that the
controller can be combined with a legacy IDE controller. The previous code
was incorrectly assuming that none of the other BARs would be
memory-based. This fixed code finds the correct BAR by index.