Fix e100 use of command interrupt bit.
The goal was to reap e100 commands every half ring cycle to allow better
throughput. Turns out it is also necessary to work around QEMU's bug where
it bails after processing 16 commands. Alas, the solution was off by one
since the free count is decremented before it is checked. So, if there are
15 free commands (i.e. 16 owned by the hardware), make sure the 16th has
the interrupt bit set so that they can be reaped in case QEMU goes out to
lunch.

Considered just setting the interrupt bit every 16th command, but that
would generate more interrupts than necessary. For example, two commands
are submitted, the first is the 16th command and the second will have the
suspend bit. Two interrupts get generated, when only the suspend bit was
really needed.