Handle TCP socket state initialized in process packet.
NetFindSocket is not synchronized with the failure of a connect operation,
so an incoming packet might find a socket, but by the time it gets to
holding the lock the connect has failed, it's only locally bound, and the
state is back at initialized. Simply drop the packet in this case.

Also added an assert that the socket should be active in a tree.