Fix Raspberry Pi UEFI time functions to fail.
The Raspberry Pi UEFI runtime time routines were returning EFI_SUCCESS
rather than failing. This was causing the boot loader to report a bogus
boot time due to an uninitialized time structure. This fixes those
routines to return EFI_UNSUPPORTED. In the kernel, the change makes sure
that if the SetTime runtime service is not present or is unsupported, that
STATUS_NO_SUCH_DEVICE is returned to KeSysteSystemTime so the entire
system call does not fail.