Synchronize PCNet32 register access.
Reading and writing to the CSR/BCR PCNet registers requires writing to an
address port and then reading from or writing to a data port. It rings a
bell that it was on my TODO list to synchronize these, but I guess I
didn't go back and do it before checking in the driver. The unsychronized
nature of this was causing the PCNet driver/device to hang, likely because
one of the register had been scrozzed.