Fix leaked page cache entries.
After a page cache entry is evicted from its file object's tree and placed
on a local destroy list, it is cleaned to make sure the metrics are
correct. Evict requests that the cleaned page cache entry is not moved to
the clean list, but clean was always removing it from its current list. So
if the evicted cache entry was actually dirty, clean was removing it from
the local destroy list. This would orphan the evicted page cache entry,
leaking the page cache entry and keeping its file object open.