Fix for getaddrinfo access violation.
The recent change that checks to see if a name server is a subdomain of
the original name to be looked up forgot to check if the name server had a
name. The original name servers that come from the kernel are just pure
addresses. Thanks to Vasco for tracking this one down.