Fix flushing file objects for a specific device.
IopFlushFileObjects takes an optional DeviceId parameter. This is used
when unmounting a volume to make sure all the files on the volume are
flushed before it is torn down. It's important to note that there may be
more than one file that has the given device ID, requiring the flush
routine to iterate through the dirty list file object list and clean all
file objects that match on the supplied device ID. A regression modified
the routine's behavior to exit as soon as one file object was flushed.