Remove shared memory object Volume0 dependency.
For long since forgotten reasons, shared memory objects were backed by
temp files in /Volume/Volume0/temp. This obviously does not work if
Volume 0 is read only, or small in size. This change converts the shared
memory objects to be backed by the page file, if one is present.

Implementing this required educating the page cache that some file objects
actually require an extra I/O flag to save file contents to a backing
store. So, the concept of a "hard flush" was added. If a page cache entry
tied to a shared memory object needs to be removed due to tight memory, a
hard flush will be performed and the shared memory object will save the
page to the page file. If there is no page file, then the hard flush will
fail and that page cache entry will remain in the cache to service shared
memory object operations. Shared memory objects fake success on normal
flush requests, but the page cache also records if a page cache entry was
ever dirty in order to detect that a hard flush may be needed.

This change also removes another related TODO with shared memory objects.
Shared memory objects were opening their backing images on Volume0 "from
kernel mode" on behalf of user mode. Yikes. That is gone now.