Further improvements to the logic in chown.
Now the new style which allows user names with dots in it takes
precedence. If a user name containing a dot or multiple dots is provided
on the command line and it exists in the system, then it is considered
as such, instead of trying to parse the group name.

If the input contains a colon, then we skip checking whether the string
represents an existing user, since we know for sure that a colon isn't a
valid user name character. We use the colon as separator as before.

If the input contains one or more dots, no colons and isn't a valid user
name then we use the last dot as separator.