Removed reverse iteration in Chalk.
Reverse iteration was causing ranges to be counterintuitive. The
expression "abcd"[1...-20] comes out to "a", rather than the empty string.
The uglier one is "a"[1...-1], which comes out to "a", rather than the
empty string. This change removes support for reverse ranges.

Also added a List.contains(obj), and fixed swapped arguments in
String.startsWith() and .endsWith().