Added support for function declarations in Chalk.
The lack of function declarations meant there was a forced order to
functions in a file. It also prevented indirect recurdion, ie function a
calls function b and function b calls function a.

This change adds support for function declarations. If left undefined,
the function symbol is equivalent to a null object (which will explode if
called).

Also made the module variable invisible in a "from xxx import yyy"
statement. The problem was that multiple "from xxx import ..." statements
were causing compile errors because the xxx module variable was already
defined.