Fix TCP recv hang caused by peeking.
If there are multiple receive segments when a recv call is made with the
PEEK flag set, the ExpectedSequence is not updated properly after the
first segment. This causes that loop to think it encountered a hole, and
clear the POLL_EVENT_IN. There was no hole however, so a subsequent call
to recv will hang waiting for an IN event that will never happen (because
the data's already there). This was causing wget to hang.