Added Chalk Core.importAllSymbols implementation.
With the addition of importAllSymbols, the statement "from some_module
import *;" works.

I'm still not 100% on the design here. Should symbols imported be visible
in the global namespace such that anyone that imports that module gets all
its imports as well? Or should there be a distinction between the public
facing namespace of a module and its more private global namespace?