Improved exception detection in Chalk.
The interpreter was running into issues if an exception occurred and was
caught in the same function. The check to see "did the frame count change"
is not sufficient to determine if an exception occurred. This change
switches it to "did the try count change", which it will when an exception
is dispatched.

There were also issues with stack adjustment after a foreign function
completed if the foreign function left extra stuff on the stack, or
reallocated the stack while running.

Also fixed a bug where all the globals were disappearing in an interactive
interpreter session if there was ever a compile error. This was because
the module being compiled got deleted if compilation failed. This change
makes that module get deleted only if it wasn't previously there, so the
interactive interpreter can persist and keep its globals.