Be more forceful about creating symlinks in msetup.
The new init script to set the hostname causes symlinks to show up in two
sources pointed at the same destination. Creating a symlink where
something previously existed was a failure. This change attempts to unlink
whatever was there for more forceful symlink creation, which is what's
done for regular files.