More fixes for NT version of SwOpen.
If FILE_APPEND_DATA is OR'd with GENERIC_WRITE, the write takes precedence.
This change fixes the CreateFile access flags for the append case and
properly handles the write-only and read-write mode flags.

This change also fixes the translation mode. CreateFile defaults to open
files in binary mode, but swiss.exe expects the default to be text mode on
Windows. If binary mode is not explicitly requested in the open flags, set
O_TEXT when opening the file descriptor with _open_osfhandle.