Fix shell :? and ? parameter expansion.
If :? or ? parameter expansion fail, an error is supposed to be printed to
standard error and the shell should exit if it is non-interactive. Swiss
had been printing to standard out and had not been setting a non-zero
return value. It was exiting, but not with an error value as it should
have been. This caused a third-party configure script to fail when it
should not have.

Tracing the exit path caught another bug in subshell group execution. If
node execution returns FALSE, the child forked for subshell execution would
actually return up the call stack, rather than calling exit(). This change
fixes that as well.