Used a fixed page file size during automation.
This is an attempt at fixing Boren, which with the change to GCC-6 has
recently started getting stuck while building GCC. It's stuck because it's
got 512MB of RAM, and so its 1GB page file is too small for make -j2 with
the right combination of files. Denny (the BeagleBone Black) will likely
be in the same situation.