Always keep latest nightlies.
Still haven't managed to get far enough to test this script, but it
occurred to me that if something like ARMv6 is way behind, the cleanup
might delete its build. Prevent this from happening by explicitly keeping
whatever latest is pointing to. This will need to be updated if new
architectures come online.