Turn off IPv6 implicit DNS translation.
If AF_UNSPEC is passed in to the hints of getaddrinfo, don't try for an
IPv6 translation. IPv6 support in the kernel is not yet done, so doing
this may lead to IPv6-only name servers, which causes the entire DNS
translation to fail. When we get IPv6 online, this small snippet can be
removed.