Add DW_AT_Specification support to DWARF parser.
A lot of the globals were no longer showing up in the new compiler (like
ImImportTable) because they were declared in the same translation unit as
they were defined. GCC was emitting a declaration (which we ignore), and
then emitting a definition that used the Specification attribute to point
at the declaration.

This change supports that by following Specification attributes behind the
scenes whenever at attribute is requested.