Fix swiss CTRL+C behavior on Windows.
When running a shell script in the current context, the interactive shell
flag is temporarily removed. This is bad if a subshell gets executed while
the flag was disabled. That subshell modifies the SIGINT signal disposition
from trap to default. It then attempts to restore the disposition to that of
the parent (should be trap), but the parent shell still doesn't have the
interactive bit set. As a result, SIGINT's disposition remains 'default'.
This causes CTRL+C to kill swiss.

As part of this change, ENABLE_PROCESSD_INPUT was disabled from console
raw mode. This prevents Windows from handling CTRL+C and just gives the
raw CTRL+C key strokes to ShReadLine, resulting in a ^C being printed to
the output stream.