Add a setenv.sh script for convenience.
This new script simply attempts to set up the build environment variables
to sane values without having to understand much of what's going on. It
also checks to make sure the tools are accessible.

If run on Minoca, it tries to match the native architecture. On other
OSes, it just defaults to x86 unless already set.