Ignore other processes' file locks on unlock.
A process can only unlock file locks that it owns and should not touch
locks created by other processes. This change modifies the unlock behavior
to silently skip locks owned by other processes, rather than ASSERT and
fail. This is not clear in the POSIX specification, but generally makes
sense and is the behavior assumed by third-party applications.