Minor fixes to the build infrastructure.
Build packet failed because SRCROOT was not exported.

SSHD seemed to fail to start up because the passwd file got overwritten
during the add_files step, and opkg didn't rerun installation for openssh
because it thought it was already done. The fix was to not recopy skel to
apps if it has already happened.

I also wanted to start saving the logs of the autoclient.