Chalk compiler stack tracking fixes.
The Chalk compiler was not accurately tracking stack usage. In the case of
an indirect call, it was failing to account for the popping of the
arguments that the indirect call does. In the case of a return, it was
failing to account for the popping a return does. This causes access to
other locals below the erroneous tracking to grab the wrong variable.