Several improvements to Chalk.
1. Prevent the Chalk app from permuting arguments, as it interferes with
bundling and script arguments.
2. Expression statements entered in the interactive interpreter now result
in printing to standard out, making the interactive interpreter a lot more
fun.
3. Exception detection needs to look at both the try count and frame
count. In the C code with no try counts an exception might get dispatched
and run, and without checking the frame count the C code doing a call
might acesss a no-longer-present stack frame.
4. The bundle code was using fseeko64. But unlike lseek, fseeko64 returns
0 on success, not the file position. So we need ftello64. We'll see if
this compiles on all the OSes.