Try to improve test -ef on Windows.
Windows always reports an st_ino of 0, so test -ef was reporting that all
files are equal. This caused postgresql to fail to build on Windows
because it thought it was being built from its source directory, so it
never ran prep_buildtree.

If the st_ino is zero, match additional fields in the stat structure, like
mtime, ctime, size, and mode.