Avoid deleting latests when uploading.
We use uniq -u to print out all lines between $latests and $stale that are
not repeated. Unfortunately if an entry is in $latests but not $stale
that causes it to get deleted. Put $latests on the command line twice to
ensure that those entires are always repeated (and thus removed by uniq
-u).