Enable broadcast receive support for IPv4.
If a packet arrives with either the global IPv4 broadcast address or the
local subnet's broadcast address, then it is to be delivered to all sockets
that match. There are two main changes to enable this behavior. A network
(NET_NETWORK_ENTRY) needs to have a GetAddressType routine to classify an
incoming packet's destination address as either broadcast or unicast. And
NetFindSocket needs to be able to be called multiple times until it no
longer finds any sockets that match the packet.

A note on finding multiple sockets: as the sockets are currently stored in
red black trees, it isn't the prettiest bit of code to iterate over the
tree in order to get all unbound, locally bound and fully bound sockets
that match an address. A hash table for each binding type that has
separate chains for each socket that compares to the same address would
greatly improve the look of this.