Clear IO_FLAG_SERVICING_FAULT after use.
After increasing the page file size from 1GB to 2GB, Boren hit a strange
assert where HardPageFaults > PageFaults. We theorized this happened
because the hard page fault accounting was double-counting: one for the
file miss, and one for the disk miss.

Clearing the flag after the first time isn't perfect because a fault that
missed at the file layer but hit at the disk layer would get counted as a
hard fault.