Allow binding sockets to broadcast addresses.
The network layer should be able to classify an address before searching
for a link to which it can be bound. IPv4, unfortunately, cannot detect a
subnet broadcast address without the link address entry. So this change
pipes the NET_NETWORK_ENTRY into NetFindEntryForAddress in order to
classify an address.