Fix Chalk stack accounting after try blocks.
The Chalk compiler was getting stack offsets wrong after a try block. This
is because 1) the exception pushed on by the interpreter was not being
accounted for, and 2) the double pop needed by the control flow between
the except and the finally was causing the stack slot accounting to be off
by one. (This is needed because except jumps to finally, skipping the
scope popping code).