Chalk: revert list add change.
The recent change to make list addition, like var a = b + c, create a new
list for the result broke the setup install.ck file. It was performing
'SystemFiles += SystemFilesX86Pcat' for the non-UEFI PC image. This was
not modifying the SystemFiles list, meaning the PC/AT images did not make
it into the final pc.img.