Another attempt at the Chalk list change.
This puts the new Chalk list behavior back, and attempts to fix up
install.ck within msetup. It should now no longer rely on += modifying the
left side, as Chalk currently treats that as a reassignment.

I notice in Python for instance a += b modifies a (rather than assigning a
new object to a), and a = a + b assigns a new object to a. This might be
more intuitive behavior than what Chalk does now. The Chalk compiler
currently compiles a += b to a = a + b, so to distinguish between the two
I'd need to probably create new operator calls for all the assigns (eg
a.__addassign(b). This would actually also allow me to customize the dict
behavior for null assign to get rid of the annoying KeyError exception
that pops up if you null-assign an unpopulated key in a dict.