Chalk: changes for bundling.
Added Core.modules() to allow the user to work with the actual dictionary
of loaded modules, and dict.keys() for enumeration. Also fixed a library
path resolution bug.

I still need to figure out 1) how to reliably get the executable path on
each OS and 2) keep the chalk.exe getopts out of the way when running in
bundled mode.