Be more tolerant of unmapped regions in cache flush.
The ELF image loader remembers the range it needs to flush after
applying relocations. This may span multiple segments with an unmapped
region in between them. Change the kernel to do all the flushing it can,
even in the presence of failures. Then change the OS image call to ignore
failures. libguile seems to be the only library to have this happen so
far.