Enable I/O and mmapping of the frame buffer.
This change enables user mode to interact with the frame buffer directly.
It does this by changing the video console driver to allow open and I/O to
the frame buffer directly. It supports mmap by returning the frame
buffer physical addresses directly when a read is handed an empty
I/O buffer.

In order to support this, MM was changed to allow creation of sections
with file objects that are not cached. It turns out the CACHE_BACKED flag
turns on nearly all the right behavior (leaving non-dirty pages alone),
except for a couple places where it was dealing directly with page cache
entries. My plan is to rename the flag from PAGE_CACHE_BACKED to just
BACKED, as I think this behavior applies to any sections that are backed
by memory owned by another component (the page cache being one example).

There's a restriction placed on the video console driver that's not
spelled out super well anywhere: if the video console is going to register
itself as a non-cached file object and return I/O buffers containing its
own pages, those I/O buffer contents need to be page aligned.