Separate sockets' local send and receive addresses.
The bind operation was meant to allow a socket to choose which link to be
associated with by supplying a local address (e.g. IPv4 address). Packets
received by that socket would have to match that local address and packets
sent would use that local address as their source address. This changes
when allowing sockets to bind to the broadcast address. Sent packets should
not have the broadcast address as their source address, as the broadcast
address is not associated with any physical link. So, when binding to the
broadcast address, the address used to send packets needs to remain the
'any' wildcard address. To do this, this change differentiates a socket's
local receive address (set via bind) and its local send address (set when
binding to a non-broadcast address).