Don't overwrite binds to broadcast on connect.
If a socket is locally bound to the broadcast address (or a multicast
address), the local receive address should not be updated when it is
connected to a remote address. Generalize this and prevent the local
receive address from being updated unless the socket's binding type is
less than locally bound or invalid.