Allow raw sockets to supply a protocol of 0.
By default, a protocol of 0 is converted into the default parent protocol
number for the protocol entry found on socket creation. For example, a
protocol of 0 supplied to UDP is converted to 17. For raw sockets, there
is not default protocol number. A protocol of 0 should be left as 0, which
is actually the IPv6 HOPOPT protocol that the raw socket may be trying to
implement.