Fix symbolic link generation on production.
The command to update the latest symbolic links was improperly following
them. So ln -s 2017-02-24 nightlies/latest-x86 was creating a symbolic
link at nightlies/latest-x86/2017-02-24, rather than updating the symbolic
link that is latest-x86. The -n flag allows latest-x86 to be treated as a
normal file in this case, not a directory to be followed.