Try to be better about honoring VESA pixel format.
Before this change we were just assuming XXRRGGBB no matter what pixel
depth we got back. This change attempts to read the pixel format from the
VESA mode information. It may not be correct for 8 bit color, as that may
involve palettes, and this code assumes TrueColor (3:3:2 format) for 8
bits.