Rework ELF image symbol resolution.
DirectFB was doing something weird where it was loading multiple dynamic
libraries with the same symbols, unloading the earlier one, and then using
the second one. With our old symbol resolution code, this caused an access
violation because the first dynamic library overrode the symbols in the
second dynamic library, but then got unloaded. So when the second dynamic
library went to call its own functions, it called into garbage where the
first dynamic library should have been.

Real symbol scoping rules dictate that symbols should be searched for in
the global scope first, meaning the executable, any shared libraries
loaded initially, and any shared libraries loaded with the global flag.
Then the symbol scope for the dynamic library being loaded should be
searched, including any of its dependencies (recursively). This should
happen in a breadth-first order, not a depth first order.

The first thing this change does is to save the primary executable in a
global, for fast and definitive access to the global scope.

Then it creates new scope arrays in the LOADED_IMAGE that contains the
depth-first search order for any particular image scope.

Finally, it routes all symbol resolution, including dladdr and dlsym,
through a common code path that searches the global scope and then the
image scope.