Add the Internet Group Management Protocol (IGMP).
IGMP is the communication protocol for IPv4 multicast. It allows a host to
communicate with a router that it has joined or lieft a multicast group. This
code is not yet active as its functionality depends on reprogramming the MAC
address filters in the link layer. That is the next change to make.