Move ClInitialize to a constructor.
One of the DirectFB example apps had a constructor that called into some
libstdc++ stuff. That caused libstdc++ to initialize itself ahead of libc,
and so the C++ cerr stream didn't get properly initialized, which caused a
crash.

Move C library initialization into a constructor so that it happens before
other code that depends on it gets a chance to run.

Note: The old ClInitialize export is still there for one more build cycle,
but will be removed after that, making this a breaking change.

There's also a minor fix in there that according to the spec if
__cxa_finalize is called without an argument it should call all the exit
handlers. Unsure if that's actually used anywhere.