Move dlsym to libc_nonshared to support RTLD_NEXT.
RTLD_NEXT needs to know the shared object that the caller of dlsym is in.
One option is to use something like __builtin_return_address(0), but
that's not ideal as it's both compiler specific, and likely to break with
things like tail calls.

This change moves dlsym into libc_nonshared, so each dynamic library gets
a copy of it. That way it can call the real dlsym (now called __dlsym)
handing it an address that is unique to the shared object (&__dso_handle).

This unfortunately doesn't compile all the way because minoca.mk adds
-nostartfiles to anything built as an "so". This is necessary for drivers
but bad for regular application-level shared objects, and there's no way
to tell the difference. The next change will differentiate between
drivers and regular shared objects, and remove -nostartfiles from regular
shared objects.