Add support for multicast bind, send and receive.
This change adds the NetAddressMulticast type to netcore. Network layers
can return this as an address type when bind it searching for a local
link to match the user provided IP address, allowing the socket to bind
to a multicast address.

In order to enable sending to multicast addresses, the data link layer
needed to be invoked to translate a multicast IP address into a multicast
MAC address. Support for this was added by changing the very specific API
to get a physical broadcast address to a routine that translates a network
layer address into the appropriate physical layer address.

Lastly, receiving multicast packets on multiple sockets is now supported
by recognizing the NetAddressMulticast type on socket search.

IGMP is still not enabled. And multicast receive is not fully supported
because the NIC filters are not yet implemented.