BIOS: Disallow <= 8-bit graphics modes, use text.
This change stops looking for 8 and 4 bit video modes, mostly because I
cannot figure out how to set the palette in the BIOS. Instead in these
cases it falls back to text mode.

Also fixed a bunch of error handling bugs where we were destroying a BIOS
context but had not successfully created one.