Move multicast link search from IGMP to IPv4.
This change rearranges the layering of multicast join/leave requests a bit
in order to allow the IGMP state to be cleaned up after a link has been
removed from the global list of links. It used to be that IGMP would
search for a link when a multicast leave request arrived and this would
fail on socket clean-up if the link had disappeared. The IGMP group state
would leak. Having the IPv4 network layer find the link and pass the link
to IGMP fixes this problem and allows the IGMP structures to eventually
be destroyed and, in turn, release the lingering references on the network
link.