select should return the number of bits selected.
The C library select routine is supposed to return the number of bits
selected in total. So selecting on a single descriptor for both read and
write should return up to 2. Our implementation was just returning the
same as from poll, which is the number of unique descriptors with events.