Rework network link checksum offload flags.
This change really just renames a few fields and flags. The network link
properties used to have a checksum flags field, but that is now a generic
capabilities field. This change also clarifies that the capabilities field
is static and used to indicate which capabilities are supported, not
necessarily which are enabled.

Also, to make things a little simpler, the get/set checksum offloading API
uses the new capabilities flags. This makes it so a driver only has to keep
one set of currently enabled capabilities. Currently these only include
the checksum offloading bits, but may soon include promiscuous mode or
multicast mode bits. It wouldn't make sense to keep track of these things
in separate fields. The checksum API will just scrub the non-checksum
related bits from the supplied flags. Other APIs should do the same.

As part of this change, it appeared that the DesignWare Ethernet NIC was
not enabling receive checksum offloading due to the device's checksum
flags being set after device reset. As a result, the reset wasn't seeing
that checksum offloading should be enabled. This is now cleaned up.